Frontend refactor (#19)
* Implement LobbyModel

A LobbyModel represents a View Model entity with basic information about
a lobby. Right now, there is only a single lobby, but in the future, we
will require a collection of these models in order to properly display a
list of available lobbies.

* Separate razor view and code-behinds

* Client/Server: Switch to a common GameModel

Instead of the Blazor frontend having to ask for every single things
separately, it now keeps a GameModel that contains everything a user
needs to know. This lowers a possibility of a desync at a cost of
slightly larger JSON trafic.

* Implement SignalR Exception handling

The SignalR hub now catches common exceptions thrown by IGame and is
able to send the error message to the client, which is able to show it
to the user.

* Remove accidental ClaudiaIDE config file